Refactoring

+++ app/src/services/accounts.ts
@@ -1,44 +1,36 @@
 import { supabase } from '@/lib/supabase/supabaseClient'
 import type { PostgrestError } from '@supabase/supabase-js'
-import type { Database } from '@/lib/supabase/database.types'
 import type { AccountMembership, MembershipRole } from '@/models'
 
 export async function createAccount(name: string) {
-  const { data, error } = await supabase.rpc("create_account", { account_name: name })
+  const { data, error } = await supabase.rpc('create_account', { account_name: name })
   return { data, error }
 }
 
-export async function getAllAccounts(user_email: string): Promise<{ data: AccountMembership[] | null, error: PostgrestError | null }> {
-  const { data, error } = await supabase
-    .from('accounts_users_association')
-    .select('account_id, role, accounts(id, name)')
-    .eq('user_email', user_email)
-    .order('accounts(name)', { ascending: true })
+export async function getAllAccounts(user_email: string): Promise<{ data: AccountMembership[] | null; error: PostgrestError | null }> {
+  const { data, error } = await supabase.from('accounts_users_association').select('account_id, role, accounts(id, name)').eq('user_email', user_email).order('accounts(name)', { ascending: true })
 
-  const accountsMemberships: AccountMembership[] | null = data?.map((item) => ({
-    id: item.account_id,
-    name: item.accounts?.name || 'Unknown Account',
-    role: item.role
-  })) || null
+  const accountsMemberships: AccountMembership[] | null =
+    data?.map((item) => ({
+      id: item.account_id,
+      name: item.accounts?.name || 'Unknown Account',
+      role: item.role,
+    })) || null
 
   return { data: accountsMemberships, error }
 }
 
 export async function getAccount(id: string) {
-  const { data, error } = await supabase
-    .from('accounts')
-    .select('*')
-    .eq('id', id)
-    .single()
+  const { data, error } = await supabase.from('accounts').select('*').eq('id', id).single()
   return { data, error }
 }
 
-export async function updateAccount(id: string, values?: {name?: string, authorize_requests?: boolean}) {
+export async function updateAccount(id: string, values?: { name?: string; authorize_requests?: boolean }) {
   const { data, error } = await supabase
     .from('accounts')
     .update({
       name: values?.name,
-      authorize_requests: values?. authorize_requests
+      authorize_requests: values?.authorize_requests,
     })
     .eq('id', id)
     .select()
@@ -46,10 +38,7 @@ export async function updateAccount(id: string, values?: {name?: string, authori
 }
 
 export async function deleteAccount(id: string) {
-  const { data, error } = await supabase
-    .from('accounts')
-    .delete()
-    .eq('id', id)
+  const { data, error } = await supabase.from('accounts').delete().eq('id', id)
   return { data, error }
 }
 
@@ -59,17 +48,14 @@ export async function addAccountUser(account_id: string, user_email: string, rol
     .insert({
       account_id,
       user_email,
-      role
+      role,
     })
     .select()
   return { data, error }
 }
 
 export async function getAccountUsers(id: string) {
-  const { data, error } = await supabase
-    .from('accounts_users_association')
-    .select('role, user_email, created_at')
-    .eq('account_id', id)
+  const { data, error } = await supabase.from('accounts_users_association').select('role, user_email, created_at').eq('account_id', id)
   return { data, error }
 }
 
@@ -77,7 +63,7 @@ export async function updateAccountUser(account_id: string, user_email: string, 
   const { data, error } = await supabase
     .from('accounts_users_association')
     .update({
-      role
+      role,
     })
     .eq('account_id', account_id)
     .eq('user_email', user_email)
@@ -86,11 +72,7 @@ export async function updateAccountUser(account_id: string, user_email: string, 
 }
 
 export async function removeAccountUser(id: string, email: string) {
-  const { data, error } = await supabase
-    .from('accounts_users_association')
-    .delete()
-    .eq('account_id', id)
-    .eq('user_email', email)
+  const { data, error } = await supabase.from('accounts_users_association').delete().eq('account_id', id).eq('user_email', email)
   return { data, error }
 }
 
@@ -100,17 +82,14 @@ export async function addAccountAddress(account_id: string, address: string, nam
     .insert({
       account_id,
       address,
-      name
+      name,
     })
     .select()
   return { data, error }
 }
 
 export async function getAccountAddresses(id: string) {
-  const { data, error } = await supabase
-    .from('accounts_addresses')
-    .select('*')
-    .eq('account_id', id)
+  const { data, error } = await supabase.from('accounts_addresses').select('*').eq('account_id', id)
   return { data, error }
 }
 
@@ -118,7 +97,7 @@ export async function updateAccountAddress(id: string, address: string, name: st
   const { data, error } = await supabase
     .from('accounts_addresses')
     .update({
-      name
+      name,
     })
     .eq('id', id)
     .eq('address', address)
@@ -127,11 +106,7 @@ export async function updateAccountAddress(id: string, address: string, name: st
 }
 
 export async function removeAccountAddress(id: string, address: string) {
-  const { data, error } = await supabase
-    .from('accounts_addresses')
-    .delete()
-    .eq('account_id', id)
-    .eq('address', address)
+  const { data, error } = await supabase.from('accounts_addresses').delete().eq('account_id', id).eq('address', address)
   return { data, error }
 }
 
@@ -140,33 +115,27 @@ export async function createAccountSecret(account_id: string, name: string) {
     .from('accounts_secrets')
     .insert({
       account_id,
-      name
+      name,
     })
     .select()
   return { data, error }
 }
 
 export async function getAccountSecrets(account_id: string) {
-  const { data, error } = await supabase
-    .from('accounts_secrets')
-    .select('*')
-    .eq('account_id', account_id)
+  const { data, error } = await supabase.from('accounts_secrets').select('*').eq('account_id', account_id)
   return { data, error }
 }
 
 export async function deleteAccountSecret(account_id: string, secret: string) {
-  const { data, error } = await supabase
-    .from('accounts_secrets')
-    .delete()
-    .eq('account_id', account_id)
-    .eq('secret', secret)
+  const { data, error } = await supabase.from('accounts_secrets').delete().eq('account_id', account_id).eq('secret', secret)
   return { data, error }
 }
 
 export async function getAccountSubscription(account_id: string) {
   const { data, error } = await supabase
     .from('accounts')
-    .select(`
+    .select(
+      `
       subscription_id,
       subscription_expiration_date,
       subscription_tiers (
@@ -176,7 +145,8 @@ export async function getAccountSubscription(account_id: string) {
         allow_secondary_listings,
         allow_custom_currencies
       )
-    `)
+    `
+    )
     .eq('id', account_id)
     .single()
   return { data, error }
@@ -188,10 +158,7 @@ export async function getAccountActiveListingsAppids(account_id: string, chain: 
 }
 
 export async function getAccountChainsParameters(account_id: string) {
-  const { data, error } = await supabase
-  .from('accounts_chains_parameters')
-  .select('*')
-  .eq('account_id', account_id)
+  const { data, error } = await supabase.from('accounts_chains_parameters').select('*').eq('account_id', account_id)
   return { data, error }
 }
 
@@ -203,7 +170,7 @@ export async function createAccountChainsParameters(account_id: string, chain_id
       chain_id,
       enable_secondary,
       secondary_fee_address,
-      secondary_percentage_fee
+      secondary_percentage_fee,
     })
     .select()
   return { data, error }
@@ -224,10 +191,7 @@ export async function updateAccountChainsParameters(account_id: string, chain_id
 }
 
 export async function getAccountCurrencies(account_id: string) {
-  const { data, error } = await supabase
-    .from('accounts_currencies')
-    .select('*')
-    .eq('account_id', account_id)
+  const { data, error } = await supabase.from('accounts_currencies').select('*').eq('account_id', account_id)
   return { data, error }
 }
 
@@ -237,19 +201,14 @@ export async function addAccountCurrency(account_id: string, currency: number, c
     .insert({
       account_id,
       currency,
-      chain_id
+      chain_id,
     })
     .select()
   return { data, error }
 }
 
 export async function removeAccountCurrency(account_id: string, currency: number, chain_id: string) {
-  const { data, error } = await supabase
-    .from('accounts_currencies')
-    .delete()
-    .eq('account_id', account_id)
-    .eq('currency', currency)
-    .eq('chain_id', chain_id)
+  const { data, error } = await supabase.from('accounts_currencies').delete().eq('account_id', account_id).eq('currency', currency).eq('chain_id', chain_id)
   return { data, error }
 }
 
@@ -260,16 +219,13 @@ export async function createAccountJwtSecret(account_id: string, name: string) {
       account_id,
       name,
     })
-    .select();
-  return { data, error };
+    .select()
+  return { data, error }
 }
 
 export async function getAccountJwtSecrets(account_id: string) {
-  const { data, error } = await supabase
-    .from('accounts_secrets')
-    .select('*')
-    .eq('account_id', account_id);
-  return { data, error };
+  const { data, error } = await supabase.from('accounts_secrets').select('*').eq('account_id', account_id)
+  return { data, error }
 }
 
 /**
@@ -278,10 +234,6 @@ export async function getAccountJwtSecrets(account_id: string) {
  * @param secret_id - The secret to delete.
  */
 export async function deleteAccountJwtSecret(account_id: string, secret: string) {
-  const { data, error } = await supabase
-    .from('accounts_secrets')
-    .delete()
-    .eq('account_id', account_id)
-    .eq('secret', secret);
-  return { data, error };
+  const { data, error } = await supabase.from('accounts_secrets').delete().eq('account_id', account_id).eq('secret', secret)
+  return { data, error }
 }
